python/pyproject.toml
[tool.poetry]
name = "glossary"
version = "0.2"
description = "Python package to define technical terms"
authors = ["Ian Flores Siaca <iflores.siaca@gmail.com>", "Greg Wilson <greg@rstudio.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3.1"
textdistance = "^4.2.0"

[tool.poetry.dev-dependencies]
sphinx = "^3.1.0"
sphinxcontrib-napoleon = "^0.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"